+++ CHANGELOG.md
@@ -15,8 +15,8 @@ v1.4
  - {CSS} Modification de custom.css
  - {SQL} Ajout d'un champs use_slider (int, default: 1) dans la table extaz_informations
  - {SQL} Ajout d'un champs background (int, default: 3) dans la table extaz_informations
- - {SQL} Ajout d'un champs chat_nb_messages (int, default: 20) dans la table extaz_informations
  - {SQL} Ajout d'un champs chat_prefix (text, default: '') dans la table extaz_informations
+ - {SQL} Ajout d'un champs chat_nb_messages (int, default: 20) dans la table extaz_informations
 
  v1.3
 ----